updates to deployed app with query

+++ src/controllers/userController.js
@@ -1,37 +1,58 @@
 const express = require("express");
 const User = require("../models/userModel");
 const ErrorHandler = require("../helpers/userErrorHandler");
+const QueryMethod = require("../helpers/query")
 
-exports.updateUser = async (request, response) => {
+exports.updateUser = async (req, res) => {
   try {
-    const id = request.params.id;
+  
+    const id = req.params.userId;
+    const user = req.user; // identify the user
+    if (user.id !== id) {   // !user && 
+      return res
+        .status(401)
+        .json({ success: false, message: "unauthorized user" });
+    }
+
+
+
+
+
+
     const findUser = await User.findById(id);
-    findUser.name = request.body.name;
-    findUser.email = request.body.email;
+    findUser.name = req.body.name;
+    findUser.email = req.body.email;
     await findUser.save();
-    return response.status(200).send({
+    return res.status(200).send({
       status: true,
       message: "Account has been updated successfully",
       updatedUser: findUser,
     });
   } catch (err) {
     const error = ErrorHandler.handleErrors(err);
-    response.status(404).json({ error });
+    res.status(404).json({ error });
   }
 };
 
-exports.getUser = async (request, response) => {
+exports.getUser = async (req, res) => {
   try {
-    const id = request.params.id;
-    const findOneUser = await User.findById(id);
 
+    const id = req.params.id;
+    const user = req.user; // identify the user
+    if (user.id !== id) {   // !user && 
+      return res
+        .status(401)
+        .json({ success: false, message: "unauthorized user" });
+    }
+
+    const findOneUser = await User.findById(id);
     if (!findOneUser) {
-      return response.status(404).send({
+      return res.status(404).send({
         status: false,
         message: "User not found",
       });
     } else {
-      return response.status(200).send({
+      return res.status(200).send({
         status: true,
         message: "User found",
         User: findOneUser,
@@ -39,12 +60,12 @@ exports.getUser = async (request, response) => {
     }
   } catch (err) {
     if (err.path === "_id") {
-      return response.status(401).send({
+      return res.status(401).send({
         status: false,
         message: "Invalid ID",
       });
     } else {
-      return response.status(500).send({
+      return res.status(500).send({
         status: false,
         message: "Server Error",
       });
@@ -54,14 +75,38 @@ exports.getUser = async (request, response) => {
 
 exports.getAllUsers = async (request, response) => {
   try {
-    const findAllUsers = await User.find();
-    return response.status(200).send({
+
+    const user = request.user
+    if (user.role.includes("user")) {
+        return response.status(400).send({
+            status: false,
+            message: "only gulp admins can delete users"
+        })
+    };
+
+    let queriedUsers = new QueryMethod(User.find(), request.query)
+      .sort()
+      .filter()
+      .limit()
+      .paginate();
+    let users = await queriedUsers.query;
+    response.status(200).json({
       status: true,
       message: "Users found",
-      AllUsers: findAllUsers,
+      count: users.length,
+      allUsers: users,
     });
+
+
+
+    // const findAllUsers = await User.find();
+    // return response.status(200).send({
+    //   status: true,
+    //   message: "Users found",
+    //   AllUsers: findAllUsers,
+    // });
   } catch (err) {
-    console.log(err);
+   // console.log(err);
     return response.status(404).send({
       status: false,
       message: "No users found",
@@ -70,18 +115,34 @@ exports.getAllUsers = async (request, response) => {
 };
 
 exports.deleteUser = async (request, response) => {
-  const { id } = request.query;
-  const findUser = await User.findByIdAndDelete(id);
-  if (findUser) {
-    return response.status(200).send({
-      status: true,
-      message: "User deleted successfully",
-      deletedUser: findUser,
-    });
-  } else {
-    return response.status(404).send({
-      status: false,
-      message: "User not found",
-    });
-  }
+  try {
+
+    const user = request.user
+    // const user = await User.findById(request.headers.id);
+     if (user.role.includes("user")) {
+         return response.status(400).send({
+             status: false,
+             message: "only gulp admins can delete users"
+         })
+     }
+
+
+
+    const { id } = request.query;
+    const findUser = await User.findByIdAndDelete(id);
+    if (findUser) {
+      return response.status(200).send({
+        status: true,
+        message: "User deleted successfully",
+        deletedUser: findUser,
+      });
+    } else {
+      return response.status(404).send({
+        status: false,
+        message: "User not found",
+      });
+    }
+  } catch (error) {
+    return response.status(400).json({ error })
+  };
 };
